Fix DiagnosticListener stress test
The stress test has a variety of problems with it.  For example, it's sharing the same `list` of active listeners across all threads by accidentally capturing it into the lambda that's used with all threads, it's accidentally capturing the `i` variable that changes based on loop iteration and thus is often going to be viewed as the same value for all of the threads, etc.  These cause failures as multiple threads are mutating non-thread-safe types and as state is being read inconsistently.  I've cleaned up the test.